Small changes to RunRemoteRegistrationTests: USE_TEST_BASENAME in default for BUILD_DIR

+++ ScriptsLib/RunRemoteRegressionTests.sh
@@ -2,12 +2,20 @@
 # Run a build and set of regression tests on the given remote machine, and pull back the built logs.
 #
 ### @todo - make MACHINE a required parameter - no good default!!!
-: ${MACHINE:=lewis-Mac}
 : ${LOGIN:=lewis}
 : ${SSH_TARGET:=$LOGIN@$MACHINE}
-: ${BUILD_DIR:=Sandbox/Stroika-Remote-Build-Dir}
+: ${BUILD_DIR:=Sandbox/Stroika-Build-Dir-$USE_TEST_BASENAME}
 : ${BRANCH:=V2.1-Dev}
 : ${BUILD_CONFIGURATIONS_MAKEFILE_TARGET:=default-configurations}
+
+if [ "$USE_TEST_BASENAME" == "" ] ; then 
+	echo "ERROR: Required setting USE_TEST_BASENAME as argument to RunRemoteRegressionTest."
+	exit 1;
+fi
+if [ "$MACHINE" == "" ] ; then 
+	echo "ERROR: Required setting MACHINE as argument to RunRemoteRegressionTest."
+	exit 1;
+fi
 
 echo ssh $SSH_TARGET rm -rf $BUILD_DIR
 ssh $SSH_TARGET rm -rf $BUILD_DIR
@@ -19,10 +27,6 @@ echo ssh $SSH_TARGET "git clone git@github.com:SophistSolutions/Stroika.git $BUI
 ssh $SSH_TARGET "git clone git@github.com:SophistSolutions/Stroika.git $BUILD_DIR --branch $BRANCH"
 
 VER=`ssh $SSH_TARGET cd $BUILD_DIR && ScriptsLib/ExtractVersionInformation.sh STROIKA_VERSION FullVersionString`
-if [ "$USE_TEST_BASENAME" == "" ] ; then 
-	echo "ERROR: Require setting USE_TEST_BASENAME as argument to RunRemoteRegressionTest."
-	exit 1;
-fi
 #see RegressionTests.sh for this name, and why we require $USE_TEST_BASENAME set
 TEST_OUT_FILE=Tests/HistoricalRegressionTestResults/REGRESSION-TESTS-$USE_TEST_BASENAME-$VER-OUT.txt
 
@@ -43,7 +47,6 @@ CMD2Exec+="export USE_TEST_BASENAME=$USE_TEST_BASENAME;"
 CMD2Exec+="cd $BUILD_DIR && ScriptsLib/RunPerformanceRegressionTests.sh"
 echo ssh $SSH_TARGET $CMD2Exec
 ssh $SSH_TARGET "$CMD2Exec"
-
 
 
 echo "Fetching results:"
